Suggest to increase start --timeout on timeout
We use 10 minutes timeout which is typically fine, but when starting k8s
cluster in github actions the cluster starts in more than 10 minutes and
fail the timeout.

The error givers a hint about the --timeout flag which typical user is
unlikely to know:

    did not receive an event with the "running" status, consider increasing start --timeout

Maybe we need to add something about inspecting the guest, but I did not
have a good idea for a hint.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>